test/worker/requirements.txt
scapy<2.6
redis==4.4.4
pyjwt==2.4.0
trol~=0.5.2
snare<0.3
